requirements.txt
streamlit
python-dotenv
langchain-pinecone==0.2.0
pinecone-client
langchain-groq==0.2.1
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.19
pytz
sentence-transformers
requests
Pillow
